requirements.txt
torch>=2.0.0
python-chess>=1.999
numpy>=1.24.0
tqdm>=4.65.0
mysql-connector-python>=8.0.0
